+++ include/binary_output.hpp
@@ -1,0 +1,14 @@
+#pragma once
+#include <Arduino.h>
+
+class binary_output
+{
+private:
+    int led_pin;
+    int state_low = 0;
+    int state_high = 1;
+    unsigned long last_trigger;
+public:
+    void init(int pin , bool reversed = 0);
+    void flashing(int times,int ms = 80,double offset=0.25);
+};